Merge

+++ neural_modelling/src/neuron/plasticity/synapse_dynamics_static_impl.c
@@ -23,7 +23,7 @@ address_t synapse_dynamics_initialise(
 
 //---------------------------------------
 void synapse_dynamics_process_neuromodulator_event(
-         uint32_t time, int16_t concentration, uint32_t neuron_index,
+         uint32_t time, int32_t concentration, uint32_t neuron_index,
      uint32_t synapse_type) {
     use(time);
     use(concentration);
